requirements.txt
python-dotenv~=0.21.0
langchain-community~=0.3.24
langchain~=0.3.27
streamlit~=1.37.1
langchain-chroma==0.2.5
langchain-groq==0.3.2
